rename dux/agents to dux/invited
